Dev.1.2.0

+++ admin/plugin.php
@@ -124,29 +124,33 @@ if ($_COOKIE["admin"] != encode('admin', $pa)) {
                             $file = scandir("../plugin");
 
                             for ($i = 2; $i < count($file); $i++) {
-
-                                $json_string = file_get_contents('../plugin/' . $file[$i] . "/version.json");
-                                $row = json_decode($json_string, true);
-                                echo "<tr>";
-                                echo "</tr>";
-                                echo "<td>" . $row["key"] . "</td>";
-                                echo "<td>" . $row["name"] . "</td>";
-                                echo "<td>" . $row["version"] . "</td>";
-                                echo "<td>" . $row["author"] . "</td>";
-                                if ($row["use"] == true) {
-                                    echo "<td>启用</td>";
-                                    echo '<td><a target="_blank" href="' . '../plugin/' . $file[$i] . "/setting.php" . '"><button class="mdui-btn mdui-btn-icon mdui-ripple">   <i style="color:#6E6E6E" class="mdui-icon material-icons">&#xe3c9;</i></button></a>';
-                                    echo '<a onclick="return confirmAct1();" href="plugin-move.php?id=' . $file[$i] . '"><button class="mdui-btn mdui-btn-icon mdui-ripple"><i style="color:#6E6E6E" class="mdui-icon material-icons">&#xe644;</i></button></a>';
-                                } else {
-                                    echo "<td>禁用</td>";
-                                    echo '<td><a target="_blank" href="' . '../plugin/' . $file[$i] . "/setting.php" . '"><button class="mdui-btn mdui-btn-icon mdui-ripple">   <i style="color:#6E6E6E" class="mdui-icon material-icons">&#xe3c9;</i></button></a>';
-                                    echo '<a onclick="return confirmAct2();" href="plugin-move.php?id=' . $file[$i] . '"><button class="mdui-btn mdui-btn-icon mdui-ripple"><i style="color:#6E6E6E" class="mdui-icon material-icons">&#xe86c;</i></button></a>';
+                                if($file[$i]==".gitkeep"){
+                                    
+                                }else{
+                                    $json_string = file_get_contents('../plugin/' . $file[$i] . "/version.json");
+                                    $row = json_decode($json_string, true);
+                                    echo "<tr>";
+                                    echo "</tr>";
+                                    echo "<td>" . $row["key"] . "</td>";
+                                    echo "<td>" . $row["name"] . "</td>";
+                                    echo "<td>" . $row["version"] . "</td>";
+                                    echo "<td>" . $row["author"] . "</td>";
+                                    if ($row["use"] == true) {
+                                        echo "<td>启用</td>";
+                                        echo '<td><a target="_blank" href="' . '../plugin/' . $file[$i] . "/setting.php" . '"><button class="mdui-btn mdui-btn-icon mdui-ripple">   <i style="color:#6E6E6E" class="mdui-icon material-icons">&#xe3c9;</i></button></a>';
+                                        echo '<a onclick="return confirmAct1();" href="plugin-move.php?id=' . $file[$i] . '"><button class="mdui-btn mdui-btn-icon mdui-ripple"><i style="color:#6E6E6E" class="mdui-icon material-icons">&#xe644;</i></button></a>';
+                                    } else {
+                                        echo "<td>禁用</td>";
+                                        echo '<td><a target="_blank" href="' . '../plugin/' . $file[$i] . "/setting.php" . '"><button class="mdui-btn mdui-btn-icon mdui-ripple">   <i style="color:#6E6E6E" class="mdui-icon material-icons">&#xe3c9;</i></button></a>';
+                                        echo '<a onclick="return confirmAct2();" href="plugin-move.php?id=' . $file[$i] . '"><button class="mdui-btn mdui-btn-icon mdui-ripple"><i style="color:#6E6E6E" class="mdui-icon material-icons">&#xe86c;</i></button></a>';
+                                    }
+    
+    
+                                    echo '<a onclick="return confirmAct();" href="plugin-del.php?id=' . $file[$i] . '"><button class="mdui-btn mdui-btn-icon mdui-ripple"><i style="color:#6E6E6E" class="mdui-icon material-icons">&#xe872;</i></button></a>&nbsp;';
+    
+                                    echo "</td>";
                                 }
-
-
-                                echo '<a onclick="return confirmAct();" href="plugin-del.php?id=' . $file[$i] . '"><button class="mdui-btn mdui-btn-icon mdui-ripple"><i style="color:#6E6E6E" class="mdui-icon material-icons">&#xe872;</i></button></a>&nbsp;';
-
-                                echo "</td>";
+                         
                             }
